requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
git+https://github.com/tontech/superlibrary.git@main
pydantic
pydantic-settings
pydantic[email]
PyJWT
bcrypt
mandrill